Move query_parameters.hpp from main/ to common/
QueryParameters is a tiny header-only value type - two small enums and a
struct wrapping them - depending only on common/common.hpp. It has no real
tie to main/, yet lived there while being pulled into low-level common/
headers (statement_type.hpp, batched_data_collection.hpp) by value, forcing
a layering back-edge that a forward declaration can't remove.

Relocating it to common/ - the layer it belongs to - removes the back-edge
at the source for every includer. Pure relocation: no TU's include closure
changes, so total_tu_include_cost is unchanged; the win is layering.

Removes 2 common -> main header back-edges (statement_type,
batched_data_collection).

common/ header -> main/ edges
before: 13
after:  11
