+++ src/include/duckdb/common/query_parameters.hpp
@@ -1,7 +1,7 @@
 //===----------------------------------------------------------------------===//
 //                         DuckDB
 //
-// duckdb/main/query_parameters.hpp
+// duckdb/common/query_parameters.hpp
 //
 //
 //===----------------------------------------------------------------------===//
